Moved email chunk based stream buffer to own class
The stream buffer used in the email subsystem will have value outside of
emailing. It has been moving into it's own class to be repurposed for
other things.
